# docs/openapi.yaml
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/cursos/{curso}": {
            "get": {
                "operationId": "cursos.show",
                "tags": [
                    "Curso"
                ],
                "parameters": [
                    {
                        "name": "curso",
                        "in": "path",
                        "required": true,
                        "description": "The curso ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`CursoResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/CursoResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "profesor",
                                                        "categorias"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "cursos.update",
                "tags": [
                    "Curso"
                ],
                "parameters": [
                    {
                        "name": "curso",
                        "in": "path",
                        "required": true,
                        "description": "The curso ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCursoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CursoResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/CursoResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "profesor",
                                                        "categorias"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cursos.destroy",
                "tags": [
                    "Curso"
                ],
                "parameters": [
                    {
                        "name": "curso",
                        "in": "path",
                        "required": true,
                        "description": "The curso ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/matriculas/{matricula}": {
            "get": {
                "operationId": "matriculas.show",
                "tags": [
                    "Matricula"
                ],
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "The matricula ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`MatriculaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/MatriculaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "estudiante",
                                                        "curso"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "matriculas.update",
                "tags": [
                    "Matricula"
                ],
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "The matricula ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMatriculaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`MatriculaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MatriculaResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "matriculas.destroy",
                "tags": [
                    "Matricula"
                ],
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "The matricula ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CategoriaResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "nombre"
                ],
                "title": "CategoriaResource"
            },
            "CursoResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "creditos": {
                        "type": "integer"
                    },
                    "profesor": {
                        "$ref": "#/components/schemas/ProfesorResource"
                    },
                    "categorias": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoriaResource"
                        }
                    }
                },
                "required": [
                    "id",
                    "nombre",
                    "creditos"
                ],
                "title": "CursoResource"
            },
            "EstudianteResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "carne": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "nombre",
                    "email",
                    "carne"
                ],
                "title": "EstudianteResource"
            },
            "MatriculaResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nota": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fecha_matricula": {
                        "type": "string"
                    },
                    "estudiante": {
                        "$ref": "#/components/schemas/EstudianteResource"
                    },
                    "curso": {
                        "$ref": "#/components/schemas/CursoResource"
                    }
                },
                "required": [
                    "id",
                    "nota",
                    "fecha_matricula"
                ],
                "title": "MatriculaResource"
            },
            "ProfesorResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "correo": {
                        "type": "string"
                    },
                    "especialidad": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nombre",
                    "correo",
                    "especialidad"
                ],
                "title": "ProfesorResource"
            },
...
            "UpdateCursoRequest": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 150
                    },
                    "creditos": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 6
                    },
                    "profesor_id": {
                        "type": "integer"
                    },
                    "categoria_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "minItems": 1
                    }
                },
                "title": "UpdateCursoRequest"
            },
            "UpdateMatriculaRequest": {
                "type": "object",
                "properties": {
                    "nota": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 99.99
                    }
                },
                "required": [
                    "nota"
                ],
                "title": "UpdateMatriculaRequest"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}